requirements.txt
moviepy==1.0.3
transformers==4.38.2
torch==2.6.0
diffusers==0.27.2
soundfile==0.12.1
Pillow==10.2.0
imageio[ffmpeg]==2.34.0
numpy==1.26.4
scipy==1.12.0
accelerate==0.27.2
openai==1.14.3
python-dotenv==1.0.1
requests==2.28.2


